feature: precheck if words is in the english dictionary and use faster concurrent data structures to store BKTree
