+++ server/src/bin/bktree-worker.rs
@@ -15,7 +15,7 @@ use trieve_server::{
     operators::{
         chunk_operator::get_last_processed_from_clickhouse,
         dataset_operator::{scroll_words_from_dataset, update_dataset_last_processed_query},
-        words_operator::{BkTree, CreateBkTreeMessage},
+        typo_operator::{BkTree, CreateBkTreeMessage},
     },
 };
